+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/internal/types/BRecordType.java
@@ -30,7 +30,6 @@ import io.ballerina.runtime.api.types.Type;
 import io.ballerina.runtime.api.utils.IdentifierUtils;
 import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BString;
-import io.ballerina.runtime.internal.TypeUtils;
 import io.ballerina.runtime.internal.values.MapValue;
 import io.ballerina.runtime.internal.values.MapValueImpl;
 import io.ballerina.runtime.internal.values.ReadOnlyUtils;
@@ -114,7 +113,10 @@ public class BRecordType extends BStructureType implements RecordType {
     public <V extends Object> V getZeroValue() {
         String typeName = this.typeName;
         if (intersectionType != null) {
-            typeName = TypeUtils.getMutableType((BIntersectionType) intersectionType).getName();
+            typeName = ReadOnlyUtils.getMutableType((BIntersectionType) intersectionType).getName();
+        }
+        if (isReadOnly()) {
+            return (V) ValueCreator.createReadonlyRecordValue(this.pkg, typeName, new HashMap<>());
         }
         return (V) ValueCreator.createRecordValue(this.pkg, typeName);
     }
