+++ tests/jballerina-unit-test/src/test/resources/test-src/bala/test_projects/test_project_selectively_immutable/constructs.bal
@@ -245,3 +245,12 @@ public type Versioning record {|
 public type EmployeeMap map<Employee>;
 
 public type EmployeeTable table<Employee>;
+
+public type User record {|
+    int id = 1;
+    string username = "admin";
+|};
+
+public type UnionMap map<User|Identifier>;
+
+public type UnionTable table<User|Identifier>;
